refactor(shadcn): refactor shadcn component locations to @/components/ui/shadcn
