+++ components/ui/shadcn/collapsible.tsx
